+++ tests/__fixtures__/typescript/hardcodedCompilerOptions/src/fn.ts
@@ -1,0 +1,5 @@
+function fn(): number {
+	return Date.now();
+}
+
+export default fn;